+++ tests/jsclabel.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { JSCLabel } from '../src/index';
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
 describe('JSCLabel unit tests', () => {
@@ -16,5 +16,11 @@ describe('JSCLabel unit tests', () => {
 		var renderedString =
 			'<text><tspan x="0px" font-weight="bold" style="white-space: pre">JSCLabel test string</tspan></text>';
 		expect(container.innerHTML.indexOf(renderedString) > -1).toBeTruthy();
+	});
+
+	test('Label should be destroyed after component is unmounted.', () => {
+		const { unmount, container } = render(<JSCLabel options={options}/>);
+		unmount();
+		expect(container.innerHTML).toBe('');
 	});
 });